refactor: fix file references
